Doxygen Error Fixes and Address comments (#3)
* Doxygen Error Fixes

* Fix build warnings that get treated as error

* Address comments from Kou

- Updates License.txt
- Update cmake toolchain
- Move whereami to `vendored`
- Use string_view instead of NOLINT std::string

+++ cpp/src/arrow/flight/sql/odbc/flight_sql/accessors/binary_array_accessor_test.cc
@@ -96,7 +96,7 @@ TEST(BinaryArrayAccessor, Test_CDataType_BINARY_Truncation) {
     // It's safe to create a std::string from this data because we know it's
     // ASCII, this doesn't work with arbitrary binary data.
     ss << std::string(buffer.data(), buffer.data() + chunk_length);
-  } while (value_offset < values[0].length() && value_offset != -1);
+  } while (value_offset < static_cast<int64_t>(values[0].length()) && value_offset != -1);
 
   ASSERT_EQ(values[0], ss.str());
 }
